Find All Collection Filter DONE!
holy canola that was so stupid, biggest hiccup was the tests being set up wrong, and the foo and bar not being loaded properly which caused much headbanging to try and find non-existent bugs, however my SQL was wrong for a long time in first attempts at the problem! VERY satisfying to complete!